New (MaCE)Lenia, composable GUI components
Added the more modern Lenia and MaCELenia, and made the GUI components
that go with it. As for GUI components, they are now COMPOSABLE,
although a bit annoying with the mouse pos that must be transmitted.
Added the Slider and LabeledSlider components. Maybe some bugs remain,
but seems pretty good!