Add hide/show components fix various bugs
I added the _handle_event and _draw methods, which
is the ones that should actually be called, and not subclassed.
This is simply done to add some 'basic' code at the beginning of the
draw and handle_event methods, that makes some common tasks for all
components, like checking if the component is visible or not. This is not
great, and I should change it later to be completely automatic, and hidden.
Likely by decorating the methods automatically when the class is created,
with __init_subclass__.